Revert "new database"
This reverts commit a7fc49f4758bfa4e98aadb86dcd27f073eef7ee7.

+++ wallet/src/main/java/com/mutualfunds/mf/entity/CustomerWalletEntity.java
@@ -14,8 +14,8 @@ import jakarta.persistence.Table;
 public class CustomerWalletEntity {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private long walletId;
-	private long customerId;
+	private int walletId;
+	private int customerId;
 	private double walletBalance = 0;
 	@Nullable
 	private Date TransationDate;
